:clipboard: Finalized copying to clipboard
Copying to the clipboard (by clicking) now shows a "notification" at the bottom of the page letting the user know they copied something

The modal has also been updated so that scrolling is possible while the modal is up. This should fix the problem on mobile phones where the screen jumps arround like it has squirrels in its pants.

Also added the script to inProgress page so that the modal actually loads instead of sending you to the backup contact page