:video_game: Completely redid Konami Code
index.js and projects.js replaced with konami.js. Konami.js handles inputing the konami code and making a "game pad" visible at the bottom of the page for mobile users. Also adds a little hint in the footer to let people know to try the konami code.

Konami code also now uses 'localStoreage' to create persistence for when the konami code is activated across pages and when refreshed or revisited.

Enter is no longer required when entering the konami code