:art: Added some Konami Stuff
Added support for selecting from the dictionaries available, as well as selecting the min and max depth for the search

Also changed the way the lines appear in the svg. Now that they are slightly transparent by default, you should be able to see the motion when you cross over an edge you've already created.